Track ingested & preprocessed movie plots data with DVC

+++ move_plot_feast/feature_repo/feature_views.py
@@ -4,7 +4,7 @@ import os
 # Define entity
 movie = Entity(name="movie_id", value_type=ValueType.STRING, description="Unique movie id")
 
-# Define your data source (adjust path as needed)
+# Define your data source
 data_dir = os.path.join(os.path.dirname(__file__), "data")
 movie_features_source = FileSource(
     path=os.path.join(data_dir, "movie_features.parquet"),
@@ -17,8 +17,8 @@ movie_features_view = FeatureView(
     entities=["movie_id"],
     ttl=None,
     schema=[
-        Feature(name="tfidf_1", dtype=ValueType.FLOAT64),
-        Feature(name="tfidf_2", dtype=ValueType.FLOAT64),
+        Feature(name="tfidf_1", dtype=ValueType.FLOAT),  # Changed from FLOAT64 to FLOAT
+        Feature(name="tfidf_2", dtype=ValueType.FLOAT),  # Changed from FLOAT64 to FLOAT
         # Add more feature columns here as needed
     ],
     source=movie_features_source,
